requirements.txt
asgiref>=3.8.1
Django>=5.1.7
djangorestframework>=3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-environ>=0.12.0
django-phonenumber-field>=8.0.0
drf-yasg>=1.21.7
gunicorn>=23.0.0
packaging>=24.2
phonenumbers>=8.13.40
python-dotenv>=1.0.1
psycopg2-binary>=2.9.10
sqlparse>=0.5.3
whitenoise>=6.9.0
dj-database-url>=2.0
pillow>=10.4.0
